Let user outside of the room to see who is in the room
This commit involves changing the authentication logic (using a uuid on
the client side), and letting every client that loads the webpage to
connect to the server (instead of letting them connect only after
clicking "create/join").

The server maintains a map tracking every users in the room. Care must
be taken to ensure there is no memory leak.

In this way we are able to delete some unnecessary code (like "user
connect" event), and cleanup a little bit.